requirement.txt
deepface==0.0.92
numpy==1.26.4
tf_keras==2.16.0
pillow==10.3.0
pandas==2.2.2
streamlit==1.35.0

